Update incremental logic for ao tables
In GPDB 7+, the pg_class.relstorage column no longer exists to tell us
if a table is heap or append-only. Instead, we now have the upstream
Postgres implementation of pluggable storage which introduces table (and
other relation type) access methods exposed in catalog table pg_am which
is referenced in the new pg_class.relam column.

Authored-by: Kevin Yeap <kyeap@vmware.com>