Updated s3 plugin test to use n1-standard-2
With the updated s3 plugin chunksize for both backup and restore
(500MB) intended for stability and performance of production workloads,
the existing instance type for s3 plugin test (n1-standard-1)
isn't sufficient considering the default concurrency of 6. Bumping up
the instance type to n1-standard-2 to accomodate the new defaults.